+++ src/Transform.cpp
@@ -1,5 +1,5 @@
 #include "Transform.hpp"
 
-Transform::Transform(int type, int axis, double value) :
+Transform::Transform(int type, int axis, float value) :
     type(type), axis(axis), value(value) {
 }
